% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{bcInputome-class}
\alias{bcInputome-class}
\title{The bcInputome class}
\description{
The bcInputome class
}
\section{Slots}{

\describe{
\item{\code{rowdata}}{a data.frame, with conserved colnames bc,rvcounts.raw}

\item{\code{assays}}{a list contains inputData objects}

\item{\code{sample}}{a character describe the name of sample}

\item{\code{active_assay}}{a character describe which assay should be used in the
downstream analysis}
}}

