% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtAdjustSpk}
\alias{brtAdjustSpk}
\title{brtAdjustSpk}
\usage{
brtAdjustSpk(obj, assay = "raw")
}
\arguments{
\item{obj}{an bcinputome object}

\item{assay}{an character describe which assay will be used}
}
\value{
an bcinputome object, with counts_adj added in specific assay
}
\description{
adjust counts using spike-in counts
}
