% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtAnnCluster}
\alias{brtAnnCluster}
\title{brtAnnClusters}
\usage{
brtAnnCluster(x, feature, marker = NULL, n_marker = 3)
}
\arguments{
\item{x}{seurat obj}

\item{feature}{a feature table}

\item{marker}{markers table returned by seurat}
}
\description{
brtAnnClusters
}
