% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtBinInput}
\alias{brtBinInput}
\title{brtBinInput}
\usage{
brtBinInput(obj, n = 0, p = 0)
}
\arguments{
\item{n}{a number describe the numeric threshold}

\item{p}{a number between 0-1 describe the proportion threshold}
}
\description{
cutoff inputome matrix by given threshold and generate an bin matrix
the number above n and p * max of each row was treated as TRUE.
}
