% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtData-class}
\alias{brtData-class}
\title{The brtData class}
\description{
The brtData class
}
\section{Slots}{

\describe{
\item{\code{data}}{a list of Matrix}

\item{\code{focus}}{the default data}
}}

