% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtEstimateBc}
\alias{brtEstimateBc}
\title{brtEstimateBc}
\usage{
brtEstimateBc(obj, background = NULL, threshold = 3)
}
\arguments{
\item{obj}{an brtstarter object with data from neurons and nor-neurons}
}
\description{
brtEstimateBc
}
