% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtEstimateBcInputNoise}
\alias{brtEstimateBcInputNoise}
\title{brtEstimateBcInputNoise}
\usage{
brtEstimateBcInputNoise(obj, threshold = 3, min.counts = 10)
}
\arguments{
\item{obj}{an bcInputome obj}

\item{threshold}{an integer control the threshold of noise, threshold is
estimated as median(nc) + threshold * IQR(nc), default to be 3}

\item{min.counts}{an integer, bcs with total counts less than min.counts
were excluded}
}
\value{
an bcinputome obj, with added slot:noise in the data in selected assay
}
\description{
estimate the noise level of inputs via the expression in the
negative control regions in each slice.
}
\details{
#  steps 1. remove counts less than min.counts; 2. perform proportion
normalization; 3 calculate nc distribution; 4. determine nc threshold,
all bcs are equal; 5. replace threshold with exceed threshold; 6. for bcs with
sum counts less than min.counts, replace the threshold with the observed
proportion in counts; 7. adjust threshold with spk counts
}
