% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtExperiment-class}
\alias{brtExperiment-class}
\title{The brtExperiment class}
\description{
The brtExperiment class
}
\section{Slots}{

\describe{
\item{\code{data}}{brtData class}

\item{\code{coldata}}{data.frame describe the features of each column-element}

\item{\code{rowdata}}{data.frame describe the features of each row-element}
}}

