% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtFeaturePlot}
\alias{brtFeaturePlot}
\title{brtFeaturePlot}
\usage{
brtFeaturePlot(
  obj,
  feature,
  seuratobj,
  label = T,
  title = NA,
  cutoff = 0,
  pt.size = 0.1,
  pt.alpha = 0.5,
  label.type = "label",
  label.size = 3,
  logcolor = F,
  ...
)
}
\description{
brtFeaturePlot
}
