% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtGetCutoff}
\alias{brtGetCutoff}
\title{brtGetCutoff
In fact, the positive cutoff could be inferred form background, using threshold
such as 3*IQR, but the negative cutoff can never be inferred from background
levels using datasets with large variances}
\usage{
brtGetCutoff(
  obj,
  feature = NULL,
  signal = "Neurons",
  background = NULL,
  threshold = 3
)
}
\arguments{
\item{obj}{an brtexperiment obj}

\item{feature}{an colname of coldata}

\item{signal}{an celltype from ident column}

\item{background}{celltype/celltypes from ident column}

\item{threshold}{an integer}
}
\value{
an list contain the cutoff and the plot of data
}
\description{
brtGetCutoff
In fact, the positive cutoff could be inferred form background, using threshold
such as 3*IQR, but the negative cutoff can never be inferred from background
levels using datasets with large variances
}
