% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtGetInfection}
\alias{brtGetInfection}
\title{brtGetInfection}
\usage{
brtGetInfection(obj, tva.c, rv.c)
}
\arguments{
\item{obj}{an brtstarter obj}

\item{tva.c}{cutoff of tva}

\item{rv.c}{cutoff of rvcounts.raw}
}
\value{
A list contains:

RPT: rv proportion in tva positive cells;

SE: spread efficiency;

IE: infection efficiency,

SP: estimated spread proportion at tva+ cells,

record$IP: infection proportion at different tva counts

record$RP: rv+ proportion at different tva counts
}
\description{
The negative cutoff can not be inferred from the background noise
distribution, instead, the 0 counts of tva or rv was treated as ground truth for
the negative. Under this estimate, the infection rate of tva- cells was
n_rv+/n_tva-, which is the spread efficiency,se. se is quite important in
inferring the  initial infection efficiency of TVA+ cells, and the cutoff of
tva- cells.
}
