% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtGetNoise}
\alias{brtGetNoise}
\title{brtGetNoise}
\usage{
brtGetNoise(obj, method = "IQR", gain = 3, usespk = TRUE, clean_nc = TRUE)
}
\arguments{
\item{obj}{an brtInputome object}

\item{threshold}{an integer control the threshold of noise. Signal should
above median(noise)+threshold*mad(noise)}

\item{plot}{logic}
}
\description{
Estimate noise level of inputs using NC expression level
}
