% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{brtGraph_bulk}
\alias{brtGraph_bulk}
\title{brtGraph_bulk}
\usage{
brtGraph_bulk(obj, slot, t = 0.7, verbose = T)
}
\arguments{
\item{obj}{an brtunity obj}

\item{slot}{counts, data or normalized}

\item{t}{threshold}
}
\description{
wrap an linkcomm object, the nodes were linked when cor of provided
dataset above the threshold
}
