% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{brtGraph_sc}
\alias{brtGraph_sc}
\title{brtGraph_sc}
\usage{
brtGraph_sc(obj, cutoff = 0)
}
\arguments{
\item{cutoff}{data <= cutoff will be treated as 0}

\item{slot}{which slot to draw graph from}
}
\description{
creat an graph of local connection from brtunity object
}
