% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtGraphPlot}
\alias{brtGraphPlot}
\title{brtGraphPlot}
\usage{
brtGraphPlot(
  obj,
  slot = "sc",
  color = "source",
  shape = "state",
  cutoff = 0,
  pt.size = 2,
  layout = "kk"
)
}
\description{
brtGraphPlot
}
