% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtHeatmap_inputome}
\alias{brtHeatmap_inputome}
\title{brtHeatmap_inputome}
\usage{
brtHeatmap_inputome(
  obj,
  slot,
  logscale,
  returnlist = FALSE,
  nodedata = "rvcounts.raw",
  left_ann = "den",
  left_width = 0.2
)
}
\description{
brtHeatmap_inputome
}
