% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtHeatmap_starter}
\alias{brtHeatmap_starter}
\title{brtHeatmap_starter}
\usage{
brtHeatmap_starter(
  obj,
  colsubset,
  rowsubset,
  slot = "data",
  logcolor = F,
  rank_row = TRUE,
  rank_col = FALSE,
  annbottom = "ident",
  annright = NULL,
  anntop = NULL,
  runnormalize = F,
  use_raster = FALSE,
  ...
)
}
\arguments{
\item{obj}{an brtstarter obj}

\item{colsubset}{how to select cells}

\item{rowsubset}{how to select barcodes}

\item{slot}{counts/data/normalized}

\item{logcolor}{whether the color space be log}

\item{rank_row}{if TRUE, the row will be draw decreased}

\item{annbottom}{should be an character column of coldata}

\item{runnormalize}{will run normalize by proportion method, will be useful
visualize the proportion of barcodes within specific neurons.}

\item{rankcolumn}{if TRUE, the column will be draw decreased, forced to be FALSE
if annbottom was provided}
}
\description{
brtHeatmap_starter
}
