% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtInputome-class}
\alias{brtInputome-class}
\title{The brtInputome class}
\description{
The brtInputome class
}
\section{Slots}{

\describe{
\item{\code{rowdata}}{a data.frame, with conserved colnames name,sample,version,bc}

\item{\code{coldata}}{a data.frame, with conserved colnames name,region,ap,ci}
}}

