% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchpattern.R
\name{brtMatch_10x_bc}
\alias{brtMatch_10x_bc}
\title{brtMatch_10x_bc}
\usage{
brtMatch_10x_bc(
  path,
  rvBarcode_pattern,
  quality_cutoff = 25,
  cellBelow = 2,
  rvBelow = 2,
  blockSize = 2e+06,
  rvBarcode_max_mismatch = 0,
  rvBarcode_roi = NULL,
  rvBarcode_cutoff = 3,
  cell_cutoff = 3,
  ncore = 1,
  rvBarcode_reference = NULL,
  cell_reference = NULL,
  total = NULL
)
}
\arguments{
\item{rvBarcode_pattern}{A character or character vector}

\item{quality_cutoff}{a integer specify low quality string}

\item{cellBelow}{a integer, the read will be drooped if the number of low quality
string in cell barcode region exceeds it.}

\item{rvBelow}{a integer, the read will be drooped if the number of low quality
string in rvbarcode region(the region could be future modified by rvBarcode_roi)
exceeds it.}

\item{blockSize}{a numeric value specify the number of reads to be load in
recursive loading}

\item{rvBarcode_max_mismatch}{a integer, the pattern match were performed by
biostring::vmatchpattern and the rvBarcode_max_mismatch control the max_mismatch
parameter}

\item{ncore}{a numeric value specify the cores used in time consuming steps}

\item{rvBarcode_reference}{a character vector provide the white list of rvBarcode,
the unmatched barcode will be droped}

\item{cell_reference}{a character vector provide the white list of cell barcode,
the unmatched cell will be droped}

\item{rvBarcode_row}{a numeric vector to extract pure barcode from barcode region}

\item{seq_library}{a character,'cdna'/'10x'}
}
\value{
a tibble with colnames cell,bc counts
}
\description{
brtMatch_10x_bc will match and the counts the umi grouped by
10x barcode and barcodes provided from library made from 10x cdna. The reads2
contains the cell barcode and the reads1 contains the target gene information
}
