% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchpattern.R
\name{brtMatch_10X_target}
\alias{brtMatch_10X_target}
\title{brtMatch_10x_target}
\usage{
brtMatch_10X_target(
  path,
  pattern,
  quality_cutoff = 25,
  cellBelow = 2,
  patternBelow = 2,
  blockSize = 2e+06,
  pattern_max_mismatch = 0,
  cell_reference = NULL,
  total = NULL
)
}
\description{
return a dataframe record the umi counts of a given pattern
}
