% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchpattern.R
\name{brtMatch_input}
\alias{brtMatch_input}
\title{brtMatch_input}
\usage{
brtMatch_input(
  path,
  indexs,
  RvPatterns,
  umiPattern,
  quality_cutoff = 25,
  rvBelow = 2,
  umiBelow = 2,
  blockSize = 2e+06,
  rvBarcode_roi = NA,
  umi_roi = NA,
  index_roi = NA,
  Batch = NA_character_,
  ID = NA_character_,
  ncore = 1,
  rvBarcode_max_mismatch = 0,
  umi_max_mismatch = 0,
  rvBarcode_reference = NULL,
  total = NULL
)
}
\arguments{
\item{umiPattern}{a character vector, describe the R2 pattern of UMI}

\item{ncore}{a integer describe how many cores to be used}

\item{sampels}{a character or character vector describe the path of each sample}

\item{Rvpatterns}{a n*3 data.frame. $pattern describe the patterns of RvBarcodes,
$name describe the name ('rvbVn' or 'spkVn') of each pattern.}
}
\value{
a list contain barcode  and spike-in umi counts tibble, with colnames
id,bc,counts,name
}
\description{
match a list of patterns, typically different verisons of barcodes,and
return a brtInput class
}
