% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brtMergeRegion}
\alias{brtMergeRegion}
\title{brtMergeRegion}
\usage{
brtMergeRegion(obj, reference)
}
\arguments{
\item{reference}{a data.frame describe the previous region name
(colname old_name), and the corresponding new region name(colname new_name).}
}
\description{
brtMergeRegion
}
