% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtNormalize}
\alias{brtNormalize}
\title{brtNormalize}
\usage{
brtNormalize(obj, method, ...)
}
\arguments{
\item{method}{can be one of proportion,zscore,logNormalize}
}
\description{
brtNormalize
}
