% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brtNormalizeInput}
\alias{brtNormalizeInput}
\title{brtNormalizeInput}
\usage{
brtNormalizeInput(obj, method, assay = "raw", focus = "data", cutoff = 0)
}
\arguments{
\item{obj}{an bcinputome/brtunity obj}

\item{method}{an character describe the normalize method}

\item{assay}{an character describe which assay to use}

\item{focus}{an character describe which data to use}
}
\value{
an obj with added slot:focus_method  in the data in
selected assay
}
\description{
normalize each bc by selected method
}
\details{
1. proportion , normalize each bc to 0-1 proportion
2. scale, center and scale data by calling base::scale
}
