% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotInfection}
\alias{brtPlotInfection}
\title{brtPlotInfection
plot the results of brtGetInfection}
\usage{
brtPlotInfection(obj, tva.c, rv.c)
}
\description{
This function probably plot the most important features of
the result of brtGetInfection, especially the SP value and the correlation
between tva counts and bc+ proportion
}
\seealso{
[brtGetInfection()] for details
}
