% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotMultiJs_list}
\alias{brtPlotMultiJs_list}
\title{brtPlotMultiJs_list}
\usage{
brtPlotMultiJs_list(obj, include_hist = FALSE, ncol = 2)
}
\arguments{
\item{obj}{a list contains the results returned from brtMultiJs}

\item{include_hist}{logical, whether the distribution of shuffling shiold be
plot}

\item{ncol}{the number of col if plot distribution}
}
\description{
plot results returned from brtMultiJs, this step usually after
splitRegion, thus the following data and results should be list
}
