% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotNoise}
\alias{brtPlotNoise}
\title{brtPlotNoise}
\usage{
brtPlotNoise(
  obj,
  vln = FALSE,
  dot = FALSE,
  slice = "all",
  ncol = 3,
  usespk = TRUE
)
}
\arguments{
\item{obj}{an brtInputome object}

\item{vln}{logic, return violin plot of signal and noise in all samples}

\item{dot}{logic, return dot plot of the noise level of each barcode across
all NCs}
}
\value{
a graph
}
\description{
Plot NC noise level and signal within samples
}
