% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotStarter_proportion}
\alias{brtPlotStarter_proportion}
\title{brtPlotStarter_proportion}
\usage{
brtPlotStarter_proportion(obj)
}
\arguments{
\item{obj}{a brtstarter obj}
}
\description{
This function plots the proportion of true starters: TVA+ G+ and
unique.proportion > 0
}
