% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotUniqueBc}
\alias{brtPlotUniqueBc}
\title{brtPlotUniqueBc
plot the results of brtSetUniqueBc}
\usage{
brtPlotUniqueBc(obj, nrow = 2)
}
\arguments{
\item{obj}{a brtstaretr obj}

\item{nrow}{control the row number of plots}
}
\description{
brtPlotUniqueBc
plot the results of brtSetUniqueBc
}
\seealso{
[brtSetUniqueBc()] for details
}
