% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtRemoveBcInputNoise}
\alias{brtRemoveBcInputNoise}
\title{brtRemoveBcInputNoise}
\usage{
brtRemoveBcInputNoise(obj, assay = "raw")
}
\arguments{
\item{obj}{an bcInputome obj}

\item{assay}{an character describe which assay will be used}
}
\value{
an scInputome obj, with with added slot:data in the data in selected
 assay, the bcs with 0 total counts were removed
}
\description{
remove noise of inputs
}
