% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtSetCellState}
\alias{brtSetCellState}
\title{brtSetCellIdentity}
\usage{
brtSetCellState(obj, starter, input)
}
\arguments{
\item{starter}{logical expression indicating the starter rows}

\item{input}{logical expression indicating the input rows}
}
\description{
set state of cells. states including naive: bc negative cells,
starter: bc+ tva+ cells and input: bc+ tva- cells. The logical expression
starter and input is free to set, butshould always include tva and rvcounts.raw.
}
