% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtSetNoiseBc}
\alias{brtSetNoiseBc}
\title{brtSetNoiseBc
Estimate the noise level of bcs}
\usage{
brtSetNoiseBc(obj, background = NULL, threshold = 3, signal = "Neurons")
}
\arguments{
\item{obj}{an brtstarter object, the idents of coldata should contain Neurons
and at least one non-neuronal ident.}

\item{background}{a character vector, describe which non-neuronal cluster used
to calculate the threshold. Using all non-neuronal groups if none is provided.}

\item{threshold}{a numeric value, default to be 3}

\item{signal}{a character, describe which ident is signal, default is Neurons}
}
\description{
The data is noisy, it is important to identify noise. RV seldom
infect non-neuronal cells from neurons, thus the bcs expressing levels of
non-neuronal groups reflects the noise level of experiment. By measure the
median + threshold * IQR, the threshold of signal can be determined.
}
