% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtSetStarter}
\alias{brtSetStarter}
\title{brtSetStarter}
\usage{
brtSetStarter(obj, starter = NULL, input = NULL)
}
\arguments{
\item{starter}{an expression character}

\item{input}{an expression character}
}
\description{
brtSetStarter
}
