% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtSetUniqueBc}
\alias{brtSetUniqueBc}
\title{brtSetUniqueBc}
\usage{
brtSetUniqueBc(obj, N = 3, U = 0.3, P = 0.8)
}
\arguments{
\item{obj}{a brtstarter obj}

\item{N}{cutoff of counts}

\item{U}{cutoff of proportion within cells}

\item{P}{cutoff of proportion across cells}
}
\value{
update rowdate and coldata with useful columns
cell.p: for bc i, cell.p is max(bci)/sum(bci) and
cell: cell with max bci
cell.u: for bc i and cell j, cell.u is bci/sum(cell.u)
unique.bc, logical, which bc is unique and useful bc.
rvcounts.unique: total unique counts within one cell, useful for downstream
analysis
}
\description{
unique bc should be non-noise bc, unique virus bc, and is distinct
among starters. Thus, brtSetNoiseBc, brtSetUniqueVirus and brtSetCellState should
have been run.
}
