% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtSetUniqueVirus}
\alias{brtSetUniqueVirus}
\title{brtSetUniqueVirus
set unique bc particles based on umi counts}
\usage{
brtSetUniqueVirus(obj, S = NULL, P = 0.95)
}
\arguments{
\item{obj}{a brtstarter obj}

\item{S}{The total number of particles, default to be NULL and S will be estimated
by the observed bcs.}

\item{P}{The cutoff unique probability of each bc.}
}
\description{
Assuming that the probability of bcs being selected from virus
library is linearly correlated with the umi counts of bcs, the total virus
particles number and the unique probability of each bc could be estimated.
}
