% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{brtSeuratWraper}
\alias{brtSeuratWraper}
\title{brtSeuratWraper}
\usage{
brtSeuratWraper(seuratobj, reduction_method = "umap")
}
\arguments{
\item{seuratobj}{seurat objects}

\item{rduction_method}{'umap' or 'tsne'}
}
\value{
a metadata of seurat
}
\description{
wrap seurat object
}
\examples{
foo<-brtSeuratWrap(brt1_seurat,level=1)
}
