% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtStarter-class}
\alias{brtStarter-class}
\title{The brtStarter class}
\description{
The brtStarter class
}
\section{Slots}{

\describe{
\item{\code{rowdata}}{a data.frame, with conserved colnames name,bc,rvcounts.raw}

\item{\code{coldata}}{a data.frame, with conserved colnames name,sample,region,cell,
rvcounts.raw}
}}

