% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{brtStarter}
\alias{brtStarter}
\title{brtStarter}
\usage{
brtStarter(x, seuratobj, virus, sample, region, verbose = TRUE)
}
\arguments{
\item{x}{an tibble returned from brtMatch_10x_bc}

\item{seuratobj}{the seurat object of this sample}

\item{virus}{a brtVirus obj}

\item{sample}{sample name}

\item{source}{from which region}
}
\value{
an brtStarter obj
}
\description{
brtStarter
}
