% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{brtStarter_from_tibble}
\alias{brtStarter_from_tibble}
\title{brtStarter_from_tibble}
\usage{
brtStarter_from_tibble(
  x,
  seuratobj,
  sample,
  source,
  features_col = NULL,
  features_row = NULL,
  treat.na = 0
)
}
\arguments{
\item{x}{an tibble returned from brtMatch_10x_bc}

\item{sample}{sample name}

\item{source}{from which region}

\item{features_col}{features to add to coldata}

\item{features_row}{features to add to rowdata}

\item{treat.na}{how to treat the unmatched frame of features provided}
}
\value{
an brtStarter obj
}
\description{
brtStarter_from_tibble
}
