% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtUnique}
\alias{brtUnique}
\title{brtUnique}
\usage{
brtUnique(
  obj,
  subset,
  cutoff.p = 0.9,
  cutoff.n = 3,
  rr = NULL,
  cutoff.within = 0.3,
  ...
)
}
\arguments{
\item{subset}{an expression describe the subset of coldata}

\item{cutoff.p}{inter group cutoff}

\item{cutoff.n}{counts cutoff}

\item{rr}{character vector provide the unique barcode reference.
Unique will be False if the barcode is not one of the reference}

\item{cutoff.within}{within group cutoff}
}
\description{
brtUnique
}
