% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{brtUniqueVirus}
\alias{brtUniqueVirus}
\title{brtUniqueVirus}
\usage{
brtUniqueVirus(x, S = NULL, N = NULL, P.min = 0.95)
}
\arguments{
\item{x}{an brtvirus object}

\item{S}{an integer describe the number of infected cells}

\item{N}{an integer describe the types of barcodes, ignored if S is provided}
}
\description{
describe unique labeling rate of the brtVirus object
}
