% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtUnity-class}
\alias{brtUnity-class}
\title{The brtUnity class}
\description{
The brtUnity class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{a list contains coldata of all the coldatas of input objs.
support one starter region in one animal only.}

\item{\code{sc}}{a list of objs records single-cell connection}

\item{\code{bulk}}{a brtInputome obj record bulk-regions connection}

\item{\code{tools}}{a list record some useful objs}
}}

