% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{brtUnity}
\alias{brtUnity}
\title{brtUnity}
\usage{
brtUnity(starter = NULL, sc = NULL, bulk = NULL)
}
\arguments{
\item{sc}{a list of brtstarter obj, but from the upstream regions,
require sample and source}

\item{bulk}{a list of brtinputome obj, require sample and source}

\item{staretr}{a list of brtstarter obj, require sample and source}
}
\value{
a brtUniry obj
}
\description{
brtUnity
}
