% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{brtVirus-class}
\alias{brtVirus-class}
\title{the brtvirus class}
\description{
the brtvirus class
}
\section{Slots}{

\describe{
\item{\code{bc}}{a data.frame describe bc and counts of each bc}

\item{\code{batch}}{which batch the virus is}

\item{\code{U1000}}{the unique proportion of virus, if 1000 virus particles were sampled}
}}

