% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{brtVirus}
\alias{brtVirus}
\title{brtVirus
construct brtVirus object}
\usage{
brtVirus(x, batch)
}
\arguments{
\item{x}{a data.frame, with colnames bc describe the sequence of each barcode
and counts describe the counts of each bc}

\item{batch}{which batch the virus is}
}
\description{
brtVirus
construct brtVirus object
}
