% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{inputData-class}
\alias{inputData-class}
\title{The inputData class}
\description{
The inputData class stores the data and describe the column feature
of the data
}
\section{Slots}{

\describe{
\item{\code{data}}{a brtData obj, stores the raw counts,normalized ... matrix}

\item{\code{coldata}}{usually describe the region, rvcounts,and spk counts of each
region. brtInputome obj usually be a part of unity obj, the rowdata has
been described in the parent object.}
}}

