% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmeanexp.R
\name{logmeanexp}
\alias{logmeanexp}
\title{Mean of Numbers in Log-Scale}
\usage{
logmeanexp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Mean of Numbers in Log-Scale
}
\examples{
log(mean(exp(seq_len(3))))
logmeanexp(seq_len(3))
}
\keyword{math}
