% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{Sum of Numbers in Log-Scale}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Sum of Numbers in Log-Scale
}
\examples{
log(mean(exp(seq_len(3))))
logmeanexp(seq_len(3))
}
\keyword{math}
