% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge.counts}
\alias{merge.counts}
\title{Merge Technical Duplicates in a Count Matrix}
\usage{
\method{merge}{counts}(count.matrix, method = "average",
  extension = "_L00[1234]", log = FALSE)
}
\description{
Merge the technical duplicate that has the same sample name but different extensions
}
\examples{
# merging the 4 lanes from NextSeq RNAseq data
counts.merged = merge.counts(count.matrix, method = 'average', extension = '_L00[1234]')
}
\seealso{
\code{read.counts()}
}

