% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{mergeRegion}
\alias{mergeRegion}
\title{mergeRegion}
\usage{
mergeRegion(obj, r, split = FALSE)
}
\arguments{
\item{obj}{an brtunity object}

\item{r}{an reference data.frame with column source and target,source is the
second level region on tree and target is the region on tree above second level}

\item{split}{logic, whether the object to target regions}
}
\value{
an brtExperiment obj
}
\description{
merge regions by reference
}
