% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel.cor}
\alias{panel.cor}
\title{Display correlation on pairs() function.}
\usage{
panel.cor(x, y, digits = 4, prefix = "", cex.cor, ...)
}
\description{
Display correlation on pairs() function.
}
\seealso{
\code{pairs()} function help where this function is scraped.
}

