% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{parseRegion}
\alias{parseRegion}
\title{parseRegion}
\usage{
parseRegion(obj, CI = "top", nodedata = "rvcounts.raw")
}
\arguments{
\item{obj}{a brtInputome obj, with regions/AP/CI information within coldata,
a join region should be linked by / ,for example "PL/ILA"}

\item{by}{"structure","slice","sample"}
}
\description{
parse input region by allen-brain annotation
}
