% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.counts}
\alias{read.counts}
\title{Generate a Count Matrix}
\usage{
read.counts(path = ".", extension = ".Aligned.out.count",
  empty.remove = TRUE, log = FALSE)
}
\value{
A count matrix in class data.frame
}
\description{
Reads all gene count files in a directory and creates a count data.frame from them. Each column of the data.frame represents a file; each row represent a gene. Empty rows are removed by default.
}
\examples{
# If I want to generate a count matrix using count files with extension
# "Aligned.out.count" in directory "~/RNAseq" I would:
read.counts(path = '~/RNAseq', extension = 'Aligned.out.count')
}

