% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{splitRegion}
\alias{splitRegion}
\title{splitRegion}
\usage{
splitRegion(obj, r)
}
\arguments{
\item{obj}{an brtunity object}

\item{r}{an reference data.frame with column source and target,source is the
second level region on tree and target is the region on tree above second level}
}
\value{
a list contain brtInput objs
}
\description{
merge regions by reference
}
