% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpval.R
\name{tpval}
\alias{tpval}
\title{The P-value of a t Test Base on a t-statistic.}
\usage{
tpval(x, df, log = FALSE)
}
\arguments{
\item{x}{a t statistic}

\item{df}{degrees of freedom}

\item{log}{the probability is in log-scale}
}
\description{
The P-value of a t Test Base on a t-statistic.
}
\examples{
tpval(1, df=3)
exp(tpval(1, df=3, log=TRUE))
tpval(Inf, df=3)
tpval(0, df=3)
}
\keyword{distribution}
