% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulize.R
\name{brtPlotNoise}
\alias{brtPlotNoise}
\title{brtPlotNoise}
\usage{
brtPlotNoise(
  obj,
  feature,
  background = NULL,
  threshold = 3,
  signal = "Neurons"
)
}
\arguments{
\item{obj}{an brtstarter object, the idents of coldata should contain Neurons
and at least one non-neuronal ident.}

\item{feature}{bc or a name of numeric column of coldata. bc will override bc
column if bc column exist.}

\item{background}{a character vector, describe which non-neuronal cluster used
to calculate the threshold. Using all non-neuronal groups if none is provided.}

\item{threshold}{a numeric value, default to be 3}
}
\description{
brtPlotNoise
}
