% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{brtUnity}
\alias{brtUnity}
\title{brtUnity}
\usage{
brtUnity(starter = list, sc = NULL, bulk = NULL)
}
\arguments{
\item{starter}{a list of starter objs}

\item{sc}{a list of brtStarter objs records single cell long-range inputs}

\item{bulk}{a list of brtInputome objs records long-range bulk inputs}
}
\value{
a brtUnity obj
}
\description{
brtUnity
}
\seealso{
[brtUnity-class]
}
